memperbaiki fungsi pembagian

+++ switch.cpp
@@ -3,34 +3,41 @@ using namespace std;
 
 int bilangan1, bilangan2;
 
-void inputdata(){
+void inputdata()
+{
     cout << "Masukan bilangan pertama:";
     cin >> bilangan1;
     cout << "Masukan bilangan kedua:";
     cin >> bilangan2;
 }
 
-int penjumlahan ( int a, int b){
+int penjumlahan(int a, int b)
+{
     return a + b;
 }
 
-int pengurangan ( int a, int b){
+int pengurangan(int a, int b)
+{
     return a - b;
 }
 
-int perkalian ( int a, int b){
+int perkalian(int a, int b)
+{
     return a * b;
 }
 
-int pembagian ( int a, int b){
+float pembagian(int a, int b)
+{
     return a / b;
 }
 
-int main(){
-     
-     int pilihan;
-     
-     do{
+int main()
+{
+
+    int pilihan;
+
+    do
+    {
         cout << "kalkulator sederhana" << endl;
         cout << "====================" << endl;
         cout << "1. Penjumlahan" << endl;
@@ -42,33 +49,32 @@ int main(){
         cout << "Masukan menu pilihan" << endl;
         cin >> pilihan;
 
-
-        switch (pilihan){
-            case 1:
+        switch (pilihan)
+        {
+        case 1:
             inputdata();
             cout << "hasil penjumlahan = " << penjumlahan(bilangan1, bilangan2) << endl;
             break;
-            case 2:
+        case 2:
             inputdata();
             cout << "hasil pengurangan = " << pengurangan(bilangan1, bilangan2) << endl;
             break;
-            case 3:
+        case 3:
             inputdata();
             cout << "hasil perkalian = " << perkalian(bilangan1, bilangan2) << endl;
             break;
-            case 4:
+        case 4:
             inputdata();
             cout << "hasil pembagian = " << pembagian(bilangan1, bilangan2) << endl;
             break;
-            case 5:
+        case 5:
             break;
 
-            default:
+        default:
             cout << "pilihan tidak ada" << endl;
         }
-        
-     }
-     while (pilihan != 5);
-     system ("PAUSE");
-     system ("CLS");
+
+    } while (pilihan != 5);
+    system("PAUSE");
+    system("CLS");
 }
